Move DevDisplay and edit getErrorIcon
