--- app/src/main/java/com/ittelekom/app/ui/util/DevDisplayUtils.kt
@@ -1,49 +1,0 @@
-package com.ittelekom.app.ui.util
-
-import androidx.compose.foundation.layout.*
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.outlined.Engineering
-import androidx.compose.material.icons.outlined.Man
-import androidx.compose.material.icons.outlined.SettingsSuggest
-import androidx.compose.material3.*
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.text.style.TextAlign
-import androidx.compose.ui.unit.Dp
-import androidx.compose.ui.unit.dp
-
-@Composable
-fun DevDisplay(
-    modifier: Modifier = Modifier,
-    iconSize: Dp = 170.dp,
-    textStyle: androidx.compose.ui.text.TextStyle = MaterialTheme.typography.bodyLarge,
-    iconTint: androidx.compose.ui.graphics.Color = MaterialTheme.colorScheme.outlineVariant,
-) {
-    Box(
-        modifier = Modifier
-            .fillMaxSize()
-    ) {
-        Column(
-            modifier = modifier
-                .fillMaxSize(),
-            verticalArrangement = Arrangement.Center,
-            horizontalAlignment = Alignment.CenterHorizontally
-        ) {
-            Icon(
-                imageVector = Icons.Outlined.SettingsSuggest,
-                tint = iconTint,
-                contentDescription = "Error Icon",
-                modifier = Modifier
-                    .size(iconSize)
-            )
-            Spacer(modifier = Modifier.height(8.dp))
-            Text(
-                text = "В разработке",
-                style = textStyle,
-                textAlign = TextAlign.Center,
-                color = MaterialTheme.colorScheme.onSurface
-            )
-        }
-    }
-}
