+++ app/src/main/java/com/ittelekom/app/ui/util/ErrorDisplayUtils.kt
@@ -9,6 +9,8 @@ import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.size
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.outlined.CloudOff
+import androidx.compose.material.icons.outlined.SentimentDissatisfied
+import androidx.compose.material.icons.outlined.SettingsSuggest
 import androidx.compose.material.icons.rounded.SignalWifiConnectedNoInternet4
 import androidx.compose.material3.Button
 import androidx.compose.material3.Icon
@@ -72,6 +74,42 @@ fun ErrorDisplay(
 fun getErrorIcon(errorMessage: String?): ImageVector {
     return when (errorMessage) {
         "Нет подключения к интернету" -> Icons.Rounded.SignalWifiConnectedNoInternet4
+        "Нет доступных тарифов" -> Icons.Outlined.SentimentDissatisfied
         else -> Icons.Outlined.CloudOff
+    }
+}
+
+@Composable
+fun DevDisplay(
+    modifier: Modifier = Modifier,
+    iconSize: Dp = 170.dp,
+    textStyle: androidx.compose.ui.text.TextStyle = MaterialTheme.typography.bodyLarge,
+    iconTint: androidx.compose.ui.graphics.Color = MaterialTheme.colorScheme.outlineVariant,
+) {
+    Box(
+        modifier = Modifier
+            .fillMaxSize()
+    ) {
+        Column(
+            modifier = modifier
+                .fillMaxSize(),
+            verticalArrangement = Arrangement.Center,
+            horizontalAlignment = Alignment.CenterHorizontally
+        ) {
+            Icon(
+                imageVector = Icons.Outlined.SettingsSuggest,
+                tint = iconTint,
+                contentDescription = "Error Icon",
+                modifier = Modifier
+                    .size(iconSize)
+            )
+            Spacer(modifier = Modifier.height(8.dp))
+            Text(
+                text = "В разработке",
+                style = textStyle,
+                textAlign = TextAlign.Center,
+                color = MaterialTheme.colorScheme.onSurface
+            )
+        }
     }
 }
